Fixed bug in Dialogue asset handling
- When overwriting a previous dialogue asset, it is now loaded and all
subassets are manually destroyed while keeping the DialogueContainer
asset. This prevents the lists in DialogueContainer from being cleared
on runtime if the dialogue was already assigned to a Dialogue script in
the scene. This also means that you don't have to manually reassign the
dialogue to a person if the dialogue has been modified in the
DialogueEditor and saved.
- TLDR: Dialogues should no longer be "randomly deleted"